requirements.txt
tensorflow
keras
numpy
pandas
Pillow
matplotlib